Fixing friend selection for InviteFriends. Removing loader for EventActivity (do not use SyndAdapter anymore)

+++ app/src/main/java/com/timappweb/timapp/rest/services/WebServiceInterface.java
@@ -54,7 +54,7 @@ public interface WebServiceInterface {
     // Event invites
     @FormUrlEncoded
     @POST("PlacesInvitations/invite/{placeId}.json")
-    Call<List<UserInvitationFeedback>> sendInvite(@Path("placeId") int placeId, @Field("ids[]") List<Integer> ids);
+    Call<List<UserInvitationFeedback>> sendInvite(@Path("placeId") int placeId, @Field("ids[]") List<Long> ids);
 
     @GET("PlacesInvitations/accept/{inviteId}.json")
     Call<RestFeedback> acceptInvite(@Path("inviteId") int inviteId);
